Reorganize macros to separate those needed for extension library for better labeling of PRs
